Add bundler caching to the Dockerfile
This change caches the results of `bundler install` by performing them
in `/tmp`. Since Bundler is set to install globally, this works because
it writes the gems globally and does not change the state of `/tmp`, so
the hashed contents of the directory are always the same.

For more information, see [this post][docker-cache].

[docker-cache]: http://ilikestuffblog.com/2014/01/06/how-to-skip-bundle-install-when-deploying-a-rails-app-to-docker/